+++ tests/testthat/test-labelled.R
@@ -152,6 +152,14 @@ test_that("take labels from LHS", {
     ),
     labelled(c(1, 5), labels = c(Good = 1, Bad = 5))
   )
+
+  expect_equal(
+    vec_c(
+      labelled(1, labels = c(Good = 1)),
+      labelled(5, labels = c(Bad = 1)),
+    ),
+    labelled(c(1, 5), labels = c(Good = 1))
+  )
 })
 
 test_that("combining picks label from the left", {