+++ NEWS.md
@@ -1,5 +1,7 @@
 # haven (development version)
 
+* Fix buglet when combining `labelled()` with identical labels.
+
 # haven 2.4.0
 
 ## New features